refactor: split lint and scoped helper modules
