+++ crates/vize_atelier_sfc/src/compile/tests/define_props_regressions.rs
@@ -1,0 +1,56 @@
+use super::super::compile_sfc;
+use crate::types::SfcCompileOptions;
+use crate::{SfcParseOptions, parse_sfc};
+
+#[test]
+fn test_define_emits_quoted_update_event_in_sfc() {
+    let source = r#"<script setup lang="ts">
+const emit = defineEmits<{
+  "update:open": [value: boolean]
+}>()
+</script>
+
+<template>
+  <button @click="emit('update:open', false)">close</button>
+</template>"#;
+
+    let descriptor = parse_sfc(source, SfcParseOptions::default()).expect("Failed to parse SFC");
+    let opts = SfcCompileOptions::default();
+    let result = compile_sfc(&descriptor, opts).expect("Failed to compile SFC");
+
+    assert!(
+        result.code.contains(r#"emits: ["update:open"]"#),
+        "quoted emits keys containing ':' must be preserved:\n{}",
+        result.code
+    );
+}
+
+#[test]
+fn test_type_based_define_props_partial_destructure_keeps_template_props() {
+    let source = r#"<script setup lang="ts">
+const { a, b } = defineProps<{ label: string, a: string, b: string }>()
+</script>
+
+<template>
+  {{ label }} - {{ a }} - {{ b }}
+</template>"#;
+
+    let descriptor = parse_sfc(source, SfcParseOptions::default()).expect("Failed to parse SFC");
+    let result = compile_sfc(&descriptor, SfcCompileOptions::default()).expect("compile");
+
+    assert!(
+        result.code.contains("__props.label"),
+        "non-destructured props should remain prop accesses:\n{}",
+        result.code
+    );
+    assert!(
+        result.code.contains("__props.a") && result.code.contains("__props.b"),
+        "destructured props should remain reactive prop accesses:\n{}",
+        result.code
+    );
+    assert!(
+        !result.code.contains("_ctx.label"),
+        "type-only props must not fall back to instance context:\n{}",
+        result.code
+    );
+}
