+++ crates/vize_atelier_sfc/src/css.rs
@@ -16,6 +16,7 @@ use vize_carton::{FxHashMap, String};
 #[cfg(feature = "native")]
 mod parser;
 mod scoped;
+mod scoped_selector;
 #[cfg(test)]
 mod tests;
 mod transform;
