+++ crates/vize_vitrine/src/napi/lint.rs
@@ -15,11 +15,10 @@ use napi::bindgen_prelude::{Error, Result, Status};
 use napi_derive::napi;
 use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
 use serde_json::{Value, json};
-use std::{
-    fs,
-    sync::atomic::{AtomicUsize, Ordering},
-};
+use std::sync::atomic::{AtomicUsize, Ordering};
 use vize_carton::append;
+
+use super::lint_fix::{is_lintable_extension, lint_file_with_optional_fix, lint_source};
 
 struct PatinaRuleMetaNapi<'a> {
     name: &'a str,
@@ -307,11 +306,7 @@ pub fn lint_patina_sfc(source: String, options: Option<PatinaLintOptionsNapi>) -
         .with_locale(locale)
         .with_help_level(help_level)
         .with_enabled_rules(enabled_rules);
-    let result = if is_standalone_html_filename(&filename) {
-        linter.lint_standalone_html(&source, &filename)
-    } else {
-        linter.lint_sfc(&source, &filename)
-    };
+    let result = lint_source(&linter, &source, &filename);
     let lsp_diagnostics = LspEmitter::to_lsp_diagnostics_with_source(&result, &source);
 
     if result.diagnostics.len() != lsp_diagnostics.len() {
@@ -452,32 +447,10 @@ pub fn lint(patterns: Vec<String>, options: Option<LintOptionsNapi>) -> Result<L
     let results: Vec<_> = files
         .par_iter()
         .filter_map(|path| {
-            let mut source = match fs::read_to_string(path) {
-                Ok(s) => s,
-                Err(_) => return None,
-            };
-
-            let filename = path.to_string_lossy().to_string();
-            let initial_result = lint_source(&linter, &source, &filename);
-            let result = if should_fix
-                && let Some(fixed_source) = apply_lint_fixes(&source, &initial_result)
-                && fixed_source != source
-            {
-                match fs::write(path, fixed_source.as_bytes()) {
-                    Ok(()) => {
-                        source = fixed_source;
-                        lint_source(&linter, &source, &filename)
-                    }
-                    Err(_) => initial_result,
-                }
-            } else {
-                initial_result
-            };
-
-            error_count.fetch_add(result.error_count, Ordering::Relaxed);
-            warning_count.fetch_add(result.warning_count, Ordering::Relaxed);
-
-            Some((filename, source, result))
+            let item = lint_file_with_optional_fix(&linter, path, should_fix)?;
+            error_count.fetch_add(item.2.error_count, Ordering::Relaxed);
+            warning_count.fetch_add(item.2.warning_count, Ordering::Relaxed);
+            Some(item)
         })
         .collect();
 
@@ -529,68 +502,6 @@ pub fn lint(patterns: Vec<String>, options: Option<LintOptionsNapi>) -> Result<L
         file_count: files.len() as u32,
         time_ms: elapsed.as_secs_f64() * 1000.0,
     })
-}
-
-fn lint_source(
-    linter: &vize_patina::Linter,
-    source: &str,
-    filename: &str,
-) -> vize_patina::LintResult {
-    if is_standalone_html_filename(filename) {
-        linter.lint_standalone_html(source, filename)
-    } else {
-        linter.lint_sfc(source, filename)
-    }
-}
-
-fn apply_lint_fixes(source: &str, result: &vize_patina::LintResult) -> Option<String> {
-    let mut edits: Vec<&vize_patina::TextEdit> = result
-        .diagnostics
-        .iter()
-        .filter_map(|diagnostic| diagnostic.fix.as_ref())
-        .flat_map(|fix| fix.edits.iter())
-        .filter(|edit| {
-            let start = edit.start as usize;
-            let end = edit.end as usize;
-            start <= end
-                && end <= source.len()
-                && source.is_char_boundary(start)
-                && source.is_char_boundary(end)
-        })
-        .collect();
-
-    if edits.is_empty() {
-        return None;
-    }
-
-    edits.sort_by_key(|edit| (edit.start, edit.end));
-    let mut selected = Vec::with_capacity(edits.len());
-    let mut last_end = 0u32;
-    for edit in edits {
-        if edit.start < last_end {
-            continue;
-        }
-        last_end = edit.end;
-        selected.push(edit);
-    }
-
-    if selected.is_empty() {
-        return None;
-    }
-
-    let mut fixed = vize_carton::CompactString::from(source);
-    for edit in selected.into_iter().rev() {
-        fixed.replace_range(edit.start as usize..edit.end as usize, &edit.new_text);
-    }
-    Some(fixed.to_string())
-}
-
-fn is_standalone_html_filename(filename: &str) -> bool {
-    filename.ends_with(".html") || filename.ends_with(".htm")
-}
-
-fn is_lintable_extension(extension: &str) -> bool {
-    matches!(extension, "vue" | "html" | "htm")
 }
 
 #[cfg(test)]
