+++ docs/content/guide/cli.md
@@ -179,7 +179,6 @@ Examples:
 ```bash
 vize lint --preset essential --max-warnings 0 src
 vize lint --preset opinionated --help-level short src
-vize lint --fix src
 vize lint --cross-file --cross-file-tree src
 vize lint --strict-reactivity src
 vize lint --format ansi src
@@ -187,15 +186,6 @@ vize lint --format plain src
 vize lint --format agent src
 vize lint --format markdown src
 ```
-
-`--fix` is intentionally conservative: it applies only non-overlapping edits from diagnostics that
-already carry a fix, writes changed files, and then lints the changed source again for the final
-report. Diagnostics without fixes remain reported.
-
-Patina also understands existing ESLint suppression pragmas for matching rule names during
-migration: `eslint-disable`, `eslint-enable`, `eslint-disable-next-line`, and
-`eslint-disable-line`. For example, `<!-- eslint-disable-next-line vue/require-v-for-key -->`
-suppresses the next-line Vize diagnostic for that Vue rule.
 
 ## Check
 
@@ -379,11 +369,6 @@ vize ide zed
 `vize lsp` starts the language server directly.
 `vize ide` adds editor-specific install and management commands for the VS Code and Zed
 integrations.
-
-Do not point editor `serverPath` settings at `node_modules/.bin/vize` from the npm package. The npm
-package is for project scripts and NAPI-backed commands; editor integrations should use the Rust
-`vize` binary from a GitHub release, Nix, a local Cargo build, or the binary auto-detected/downloaded
-by the VS Code extension.
 
 ## Global Options
